+++ NEWS.md
@@ -1,3 +1,12 @@
+# ResultModelManager 0.5.9
+
+Changes:
+
+1. Changes made to make package CRAN compliant
+
+2. `PooledConnectionHandler` now uses withr::defer to automatically returned pooled connections returned with 
+`getConnection`
+
 # ResultModelManager 0.5.8
 
 Bug fixes: